requirement.txt
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=2.0.0
ta>=0.11.0
python-binance>=1.0.19